backend/requirements.txt
fastapi[all]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
reportlab
openai